requirements.txt
discord.py
flask
aiohttp
python-dotenv
